backend/requirements.txt
fastapi
uvicorn
litellm
google-generativeai
openai
anthropic
fpdf2
python-dotenv
pydantic
requests